Rename config_yaml to config_driver
Nothing requires us to use yaml for the driver config, it's just what
we're currently using. So I'm renaming this to config_driver.

Also we need to make sure to load our config_oslo module so the option
is present when get_driver_config() is called.
